+++ sdk/search/Azure.Search.Documents/CHANGELOG.md
@@ -22,7 +22,8 @@
 - Renamed methods on `SearchIndexerClient` matching "\*DataSource" to "\*DataSourceConnection".
 - Split `SearchServiceClient` into `SearchIndexClient` for managing indexes, and `SearchIndexerClient` for managing indexers, both of which are now in `Azure.Search.Documents.Indexes`.
 - `SearchClient.IndexDocuments` now throws an `AggregateException` wrapping all the `RequestFailedException`s in the batch.
-- Removed `dynamic` support from `SearchDocument` for the time being
+- Removed `dynamic` support from `SearchDocument` for the time being.
+
 
 ## 1.0.0-preview.3 (2020-05-05)